+++ wayland-client/CHANGELOG.md
@@ -2,9 +2,14 @@
 
 ## Unreleased
 
+#### Additions
+
+- Support absolute paths in `WAYLAND_DISPLAY`
+
 #### Bugfixes
 
 - Fix `EventQueue::blocking_dispatch()` not flushing the connection as it should
+- Ensure that `XDG_RUNTIME_DIR` is an absolute path before trying to use it
 
 ## 0.30.0-beta.9