+++ wayland-client/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 0.30.0-beta.10
+
 #### Additions
 
 - Support absolute paths in `WAYLAND_DISPLAY`